Refactor not to export SharedContent in stores. It is not a store and utility for SharedContents.
