+++ services/conference/src/scripts/components/footer/share/Entrance.tsx
@@ -11,7 +11,7 @@ import ScreenShareIcon from '@material-ui/icons/ScreenShare'
 import StopScreenShareIcon from '@material-ui/icons/StopScreenShare'
 import SubjectIcon from '@material-ui/icons/Subject'
 import {assert} from '@models/utils'
-import {createContentOfVideo} from '@stores/sharedContents/SharedContent'
+import {createContentOfVideo} from '@stores/sharedContents/SharedContentCreator'
 import JitsiMeetJS, {JitsiLocalTrack} from 'lib-jitsi-meet'
 import {useObserver} from 'mobx-react-lite'
 import React from 'react'
