+++ services/conference/src/scripts/components/footer/share/ShareDialog.tsx
@@ -2,7 +2,7 @@ import {useStore as useMapStore} from '@hooks/MapStore'
 import Dialog from '@material-ui/core/Dialog'
 import DialogContent from '@material-ui/core/DialogContent'
 import DialogTitle from '@material-ui/core/DialogTitle'
-import {createContentOfIframe, createContentOfText} from '@stores/sharedContents/SharedContent'
+import {createContentOfIframe, createContentOfText} from '@stores/sharedContents/SharedContentCreator'
 import sharedContents from '@stores/sharedContents/SharedContents'
 import React, {useState} from 'react'
 import {Entrance} from './Entrance'
