+++ services/conference/src/scripts/stores/participants/plugins/SharedContentsLookup.ts
@@ -1,6 +1,5 @@
 import {connection} from '@models/api'
 import {action, decorate, observable} from 'mobx'
-import {SharedContent} from '../../sharedContents/SharedContent'
 import {default as sharedContents} from '../../sharedContents/SharedContents'
 import {default as participants} from '../Participants'
 import {ParticipantStorePlugin} from './utils'
